requirements.txt
django
djangorestframework
django-webpack-loader
coreapi
psycopg2
django-rest-swagger